r/kubernetes_cluster: conditionally specifying either the service_principal or MSI blocks
Previously it was possible to use a Service Principal for an AKS Cluster and also give it a
Managed Identity. However due to a breaking change in the AKS API this is no longer possible.

Instead attempting to do this via the API creates an MSI Enabled cluster and ignores the Service
Principal that's specified - as such it's only possible to create a Service Principal or MSI cluster,
not the Service Principal cluster with a Managed Identity attached as before.